Update 423.md
Resolved issue "Provide the final function in one code block on this final card. It's a little hard to follow when every line of code is broken down one by one (student might forget the lines that came previously). Ask the reader to refer back (if they are confused) to the in depth explanations of each line after presenting the whole code together"

Now issue #77 is closed